+++ src/apis/GooglePlay.js
@@ -1,5 +1,5 @@
 const { APIWrapper } = require('../')
-const gPlay = require('google-play-scraper')
+const gPlay = require('google-play-scraper').memoized()
 
 module.exports = class GooglePlayStore extends APIWrapper {
   constructor () {
@@ -11,4 +11,9 @@ module.exports = class GooglePlayStore extends APIWrapper {
   async searchApp (query) {
     return gPlay.search({ term: query, num: 10, fullDetail: true })
   }
+
+  async searchDev (query) {
+    return gPlay.developer({ devId: query, fullDetail: true, num: 10 })
+  }
 }
+
